refacto: move commands in good folders
