+++ bot/commands/commands/admin.commands.ts
@@ -1,8 +1,8 @@
-import { handleChallenge } from './commands/challenge.command';
-import { handleProposeDates } from './commands/propose-dates.command';
-import { handleScheduleChallenge } from './commands/schedule.command';
-import { handleSubmitResult } from './commands/submit-result.command';
-import { handleViewChallenge } from './commands/view-challenge.command';
+import { handleChallenge } from './commands/challenge/challenge.command';
+import { handleProposeDates } from './commands/challenge/propose-dates.command';
+import { handleScheduleChallenge } from './commands/challenge/schedule.command';
+import { handleSubmitResult } from './commands/challenge/submit-result.command';
+import { handleViewChallenge } from './commands/challenge/view-challenge.command';
 import { handleCheckTimeouts } from './commands/check-timeouts.command';
 import { handleForceResult } from './commands/force-result.command';
 import { handleForfeit } from './commands/forfeit.command';
