+++ bot/commands/commands/challenge.commands.ts
@@ -1,6 +1,6 @@
-import { handleChallenge } from './commands/challenge.command';
-import { handleProposeDates } from './commands/propose-dates.command';
-import { handleScheduleChallenge } from './commands/schedule.command';
-import { handleSubmitResult } from './commands/submit-result.command';
+import { handleChallenge } from './commands/challenge/challenge.command';
+import { handleProposeDates } from './commands/challenge/propose-dates.command';
+import { handleScheduleChallenge } from './commands/challenge/schedule.command';
+import { handleSubmitResult } from './commands/challenge/submit-result.command';
 
 export { handleChallenge, handleProposeDates, handleScheduleChallenge, handleSubmitResult };
