+++ bot/commands/commands/commands/challenge/propose-dates.command.ts
@@ -1,14 +1,14 @@
 import { ChatInputCommandInteraction } from 'discord.js';
-import { logger } from '../../../utils/logger.utils';
-import { ChallengeService } from '../../../services/tournament/challenge.services';
+import { logger } from '../../../../utils/logger.utils';
+import { ChallengeService } from '../../../../services/tournament/challenge.services';
 import {
   createErrorEmbed,
   createChallengeEmbed,
   formatTimestamp,
   StatusIcons,
-} from '../../../helpers/message.helpers';
-import { TeamTournament } from '../../../database/models';
-import { ITeamMember } from '../../../database/models/team.model';
+} from '../../../../helpers/message.helpers';
+import { TeamTournament } from '../../../../database/models';
+import { ITeamMember } from '../../../../database/models/team.model';
 
 const challengeService = new ChallengeService();
